regex, responsive, correction bug, ajout admin

+++ backend/controllers/comment.js
@@ -8,10 +8,10 @@ exports.create = (req, res, next) => {
         postId: req.body.postId,
     };
     let sql = `UPDATE post SET post.comment = post.comment + 1 WHERE postId=?;`;
-    pool.execute(sql,[req.body.postId], function (err, result) {
+    pool.execute(sql, [req.body.postId], function (err, result) {
         if (err) res.status(400).json({ err });
         let sql = `INSERT INTO comment (comment, authorId, postId) VALUES (?,?,?);`;
-        pool.execute(sql,[post.comment,post.authorId,post.postId], function (err, result) {
+        pool.execute(sql, [post.comment, post.authorId, post.postId], function (err, result) {
             if (err) throw err;
             console.log(result)
             res.status(201).json({ message: `Commentaire ajouté` });
@@ -28,14 +28,23 @@ exports.getAll = (req, res, next) => {
 }
 
 exports.delete = (req, res, next) => {
-    let sql2 = `DELETE from comment WHERE idComment=?;`;
-    pool.execute(sql2,[req.params.commentId], function (err, result) {
-        if (err) res.status(400).json({ err });
-        let sql = `UPDATE post SET post.comment = post.comment - 1 WHERE postId=?;`;
-        pool.execute(sql,[req.params.postId], function (err, result) {
+    let sql3 = `SELECT * from comment WHERE idComment=?`;
+    pool.execute(sql3, [req.params.commentId], function (err, result) {
+        if (result[0].authorId == req.body.userId || req.body.admin == true) {
             if (err) res.status(400).json({ err });
-            res.status(200).json(result)
-        });
-    })
+            let sql2 = `DELETE from comment WHERE idComment=?;`;
+            pool.execute(sql2, [req.params.commentId], function (err, result) {
+                if (err) res.status(400).json({ err });
+                let sql = `UPDATE post SET post.comment = post.comment - 1 WHERE postId=?;`;
+                pool.execute(sql, [req.params.postId], function (err, result) {
+                    if (err) res.status(400).json({ err });
+                    res.status(200).json(result)
+                });
+            })
+        } else {
+            res.status(400).json({ message : "Bien essayé petit malin!" });   
+        }
+    });
+
 }
